+++ qt/src/document_index_view.cpp
@@ -148,7 +148,7 @@ void DocumentIndexView::selectItem(DocumentIndexType type, unsigned int index, b
 	if (type == eDocIndex_Account)
 	{
 		itemToSelect = m_pModel->index(0, 0); // first level
-		itemToSelect = itemToSelect.child(index, 0);
+		itemToSelect = itemToSelect.child((int)index, 0);
 	}
 	
 	if (itemToSelect.isValid())
